evaluation: Added `ClassificationMetricsTF` class
